Removed dependency on ROM-internal arrays
I looked at the base stats array and determined that, especially if I
limited it to just the data I needed, that it wouldn't be too bad a
thing to just include it and the other two arrays I need in my multiboot
image rather than reference the ones already located in the game ROM.
This way, we get back compatibility with all previously-compatible ROMs,
and not just ones that I have dumped.

New issue: Deoxys's base stats are actually different per-game, though,
so a special case will have to be written for that.