Fixed Shedinja hash collision
A Shedinja will always have the same IVs, personality value, and
original trainer as the Nincada that generated it, meaning that it is
guaranteed to have the same hash as the Ninjask that the Nincada evolved
into. To cirvument this, there is now a boolean field in the hash
determiner that is set if and only if the Pokémon is a Shedinja. This
allows the Ninjask to be considered the same Pokémon as the Nincada it
evolved from, but for the Shedinja to be considered a new Pokémon.

Because the hash determiner has changed, all old hashes are now invalid.